Minor edits.
This is related to https://github.com/GoogleChrome/sw-helpers/pull/199, which I should have looked at when it came through.

I don't understand what this promise returns. It should say something like 'resolves to void', 'resolves to a boolean value indicating...', or 'resolves to a reference to the new `Widget` instance, for example.