Add fetch fallback option for generate SW
This provides a fallback mechanism when both network and cache is not available

All of the built-in caching strategies reject in a consistent manner
when there's a network failure and/or a cache miss. This promotes the
pattern of setting a global "catch" handler to deal with any failures in
a single handler function.

Reference:
https://developers.google.com/web/tools/workbox/guides/advanced-recipes#comprehensive_fallbacks