fix: 🐛 workbox-cli ask root of web app question
Make separator unselectable as an answer to root of web application
question when multiple child directories are present. Fix the manual
input to root of the web application question - when manual input option
is chosen a new question / input will be provided to the user to specify
the root of the web app.

✅ Closes: #2764